+++ sea-query-postgres/src/lib.rs
@@ -1,3 +1,5 @@
+#![forbid(unsafe_code)]
+
 use std::error::Error;
 
 use bytes::BytesMut;